processing/pyproject.toml
[project]
name = "processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "common_funcs",
    "duckdb>=1.3.2",
    "loguru>=0.7.3",
    "marimo>=0.14.13",
    "matplotlib>=3.10.0",
    "networkx>=3.3",
    "numpy<2.0",
    "pandas>=2.3.1",
    "pyyaml>=6.0.1",
    "python-louvain>=0.16",
    "pydash>=8.0.5",
    "scispacy>=0.5.5",
    "tqdm>=4.67.1",
    "yiutils",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.12.5",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "ty>=0.0.1a16",
]

[tool.uv.sources]
common_funcs = { path = "../src/common_funcs", editable = true }
yiutils = { path = "../src/yiutils", editable = true }

[tool.ty.rules]
# unresolved-import = "ignore"

[tool.ruff]
line-length = 79
